Break out Kicker to its own component
