+++ src/web/components/SmallHeadline.tsx
@@ -4,8 +4,8 @@ import { css, cx } from 'emotion';
 import { headline } from '@guardian/src-foundations/typography';
 import { palette } from '@guardian/src-foundations';
 
-import { PulsingDot } from '@root/src/web/components/PulsingDot';
 import { QuoteIcon } from '@root/src/web/components/QuoteIcon';
+import { Kicker } from '@root/src/web/components/Kicker';
 
 const fontStyles = (size: SmallHeadlineSize) => css`
     ${headline[size]()};
@@ -45,20 +45,6 @@ const underlinedStyles = (size: SmallHeadlineSize) => {
 
 const colourStyles = (colour: string) => css`
     color: ${colour};
-`;
-
-const kickerStyles = (colour: string) => css`
-    color: ${colour};
-    font-weight: 700;
-    margin-right: 4px;
-`;
-
-const slashStyles = css`
-    &::after {
-        content: '/';
-        display: inline-block;
-        margin-left: 4px;
-    }
 `;
 
 const linkStyles = css`
@@ -134,20 +120,5 @@ export const SmallHeadline = ({
                 {headlineString}
             </Headline>
         </h4>
-    );
-};
-
-const Kicker = ({
-    text,
-    pillar = 'news',
-    showPulsingDot,
-    showSlash = true,
-}: KickerType) => {
-    const kickerColour = palette[pillar].main;
-    return (
-        <span className={kickerStyles(kickerColour)}>
-            {showPulsingDot && <PulsingDot colour={kickerColour} />}
-            <span className={cx(showSlash && slashStyles)}>{text}</span>
-        </span>
     );
 };
